Split IPC protocol and IPC impl
