+++ settings.gradle.kts
@@ -13,6 +13,7 @@ include("web-protocol")
 include("web-adapter")
 include("webrtc-pion")
 include("webrtc-pion-ipc")
+include("webrtc-pion-ipc-protocol")
 include("common")
 include("common-protocol")
 include("service-velocity")
