+++ web-adapter/src/main/java/dev/minceraft/sonus/web/adapter/config/WebConfig.java
@@ -1,7 +1,8 @@
 package dev.minceraft.sonus.web.adapter.config;
 
 import dev.minceraft.sonus.common.config.ISubConfig;
-import dev.minceraft.sonus.web.pion.PionApi;
+import dev.minceraft.sonus.web.pion.ipc.model.BundlePolicy;
+import dev.minceraft.sonus.web.pion.ipc.model.IceServer;
 import org.jspecify.annotations.Nullable;
 import org.spongepowered.configurate.objectmapping.ConfigSerializable;
 
@@ -28,10 +29,10 @@ public class WebConfig implements ISubConfig {
             String url, @Nullable String user, @Nullable String auth
     ) {
 
-        public PionApi.IceServer create() {
-            return new PionApi.IceServer(this.url, this.user, this.auth);
+        public IceServer create() {
+            return new IceServer(this.url, this.user, this.auth);
         }
     }
 
-    public PionApi.BundlePolicy bundlePolicy = PionApi.BundlePolicy.MAX_BUNDLE;
+    public BundlePolicy bundlePolicy = BundlePolicy.MAX_BUNDLE;
 }
