+++ webrtc-pion-ipc-protocol/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    api(libs.netty.handler)
+    api(projects.commonProtocol)
+}
