+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/IpcError.java
