+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/IpcMessage.java
@@ -1,6 +1,7 @@
 package dev.minceraft.sonus.web.pion.ipc;
 // Created by booky10 in Sonus (6:19 PM 06.03.2026)
 
+import dev.minceraft.sonus.common.protocol.util.VarInt;
 import io.netty.buffer.ByteBuf;
 import org.jspecify.annotations.NullMarked;
