+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/model/BundlePolicy.java
@@ -1,0 +1,12 @@
+package dev.minceraft.sonus.web.pion.ipc.model;
+
+import org.jspecify.annotations.NullMarked;
+
+@NullMarked
+public enum BundlePolicy {
+
+    UNKNOWN,
+    BALANCED,
+    MAX_COMPAT,
+    MAX_BUNDLE,
+}
