+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/model/IceConnectionState.java
@@ -1,15 +1,16 @@
-package dev.minceraft.sonus.web.pion;
+package dev.minceraft.sonus.web.pion.ipc.model;
 // Created by booky10 in Sonus (5:16 PM 06.03.2026)
 
 import org.jspecify.annotations.NullMarked;
 
 @NullMarked
-public enum PeerConnectionState {
+public enum IceConnectionState {
 
     UNKNOWN,
     NEW,
-    CONNECTING,
+    CHECKING,
     CONNECTED,
+    COMPLETED,
     DISCONNECTED,
     FAILED,
     CLOSED,
