+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/model/PeerConnectionState.java
@@ -1,16 +1,15 @@
-package dev.minceraft.sonus.web.pion;
+package dev.minceraft.sonus.web.pion.ipc.model;
 // Created by booky10 in Sonus (5:16 PM 06.03.2026)
 
 import org.jspecify.annotations.NullMarked;
 
 @NullMarked
-public enum IceConnectionState {
+public enum PeerConnectionState {
 
     UNKNOWN,
     NEW,
-    CHECKING,
+    CONNECTING,
     CONNECTED,
-    COMPLETED,
     DISCONNECTED,
     FAILED,
     CLOSED,
