+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/pionbound/IpcApiAllocatePeer.java
@@ -2,10 +2,10 @@ package dev.minceraft.sonus.web.pion.ipc.pionbound;
 // Created by booky10 in Sonus (6:30 PM 06.03.2026)
 
 import dev.minceraft.sonus.common.protocol.util.Utf8String;
-import dev.minceraft.sonus.web.pion.PionApi;
-import dev.minceraft.sonus.web.pion.PionApi.BundlePolicy;
 import dev.minceraft.sonus.web.pion.ipc.IpcMessage;
 import dev.minceraft.sonus.web.pion.ipc.IpcTypes;
+import dev.minceraft.sonus.web.pion.ipc.model.BundlePolicy;
+import dev.minceraft.sonus.web.pion.ipc.model.IceServer;
 import io.netty.buffer.ByteBuf;
 import org.jspecify.annotations.NullMarked;
 
@@ -14,11 +14,11 @@ import java.util.List;
 @NullMarked
 public class IpcApiAllocatePeer extends IpcMessage {
 
-    private final List<PionApi.IceServer> iceServers;
+    private final List<IceServer> iceServers;
     private final BundlePolicy bundlePolicy;
     private final String id;
 
-    public IpcApiAllocatePeer(int handlerId, List<PionApi.IceServer> iceServers, BundlePolicy bundlePolicy, String id) {
+    public IpcApiAllocatePeer(int handlerId, List<IceServer> iceServers, BundlePolicy bundlePolicy, String id) {
         super(handlerId);
         this.iceServers = iceServers;
         this.bundlePolicy = bundlePolicy;
@@ -33,7 +33,7 @@ public class IpcApiAllocatePeer extends IpcMessage {
         Utf8String.write(buf, this.id);
     }
 
-    public List<PionApi.IceServer> getIceServers() {
+    public List<IceServer> getIceServers() {
         return this.iceServers;
     }
