+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/sonusbound/IpcPeerError.java
