+++ webrtc-pion-ipc-protocol/src/main/java/dev/minceraft/sonus/web/pion/ipc/sonusbound/IpcPeerOnConnectionStateChange.java
@@ -2,9 +2,9 @@ package dev.minceraft.sonus.web.pion.ipc.sonusbound;
 // Created by booky10 in Sonus (7:22 PM 06.03.2026)
 
 import dev.minceraft.sonus.common.protocol.util.VarInt;
-import dev.minceraft.sonus.web.pion.PeerConnectionState;
 import dev.minceraft.sonus.web.pion.ipc.IpcMessage;
 import dev.minceraft.sonus.web.pion.ipc.IpcTypes;
+import dev.minceraft.sonus.web.pion.ipc.model.PeerConnectionState;
 import io.netty.buffer.ByteBuf;
 import org.jspecify.annotations.NullMarked;
