+++ webrtc-pion-ipc/build.gradle.kts
@@ -1,7 +1,5 @@
 dependencies {
     api(libs.slf4j.api)
-    api(libs.netty.handler)
-
     api(projects.network)
-    api(projects.commonProtocol)
+    api(projects.webrtcPionIpcProtocol)
 }
